requirements.txt
torch>=1.11.0
transformers>=4.25.0
pybullet>=3.2.0
opencv-python>=4.5.0
tqdm>=4.64.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-learn>=0.24.0
json5
omegaconf
einops